pyproject.toml
[tool.poetry]
name = "fastapi-railway"
version = "0.1.0"
description = "A FastAPI application ready for Railway deployment"
authors = ["Alexanderdunlop <alexjamesdunlop@gmail.com>"]
packages = [
    { include = "my_agent" }
]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langgraph = "^0.2.0"
langchain_anthropic = "^0.1.0"
langchain_core = "^0.2.0"
langchain_openai = "^0.1.0"
tavily-python = "^0.3.0"
langchain_community = "^0.2.0"
pinecone = "^5.4.2"
pydantic-settings = "^2.7.0"
tiktoken = "^0.8.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 